+++ ably/ablytest/recorders.go
@@ -14,7 +14,7 @@ import (
 	"time"
 
 	"github.com/ably/ably-go/ably"
-	"github.com/ably/ably-go/ably/internal/ablyutil"
+	"github.com/ably/ably-go/ably/ablyutil"
 	"github.com/ably/ably-go/ably/proto"
 )
