+++ ably/options.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/ably/ably-go/ably/internal/ablyutil"
+	"github.com/ably/ably-go/ably/ablyutil"
 	"github.com/ably/ably-go/ably/proto"
 )
